deprecating np types

+++ regtricks/application_helpers.py
@@ -194,7 +194,7 @@ def sum_array_blocks(array, factor):
     if np.any(np.mod(factor, np.ones_like(factor))):
         raise RuntimeError("factor must be of integer values only")
 
-    factor = np.array(factor, dtype=np.int32)
+    factor = np.array(factor, dtype=int)
     if (np.mod(array.shape, factor) > 0).any(): 
         raise RuntimeError(("Factor {} must be a perfect divisor of array shape {}".
             format(factor, array.shape)))
